+++ doc/CHANGELOG.md
@@ -15,6 +15,7 @@ files use git-show or git-log.
 - actual line number highlight,
 - permament line numbers blocks,
 - a few Common User Access bindings,
+- lower pane with keybindings,
 - private mode deleted, keyboard shortcuts also don't generete signals now.
 
 ### v2.1.0 (2018, November 10):